+++ tests/components/switch/test_vultr.py
@@ -16,7 +16,7 @@ from homeassistant.components.vultr import (
 from homeassistant.const import (
     CONF_PLATFORM, CONF_NAME)
 
-from tests.components.test_vultr import VALID_CONFIG
+from tests.components.vultr.test_init import VALID_CONFIG
 from tests.common import (
     get_test_home_assistant, load_fixture)
